pyproject.toml
[project]
name = "nonebot_plugin_fishcat_ping"
version = "1.0.0"
description = "一个简单而实用的 NoneBot2 ping 插件"
authors = [
    {name = "tniany", email = "your.email@example.com"}
]
dependencies = [
    "nonebot2>=2.0.0",
    "nonebot-adapter-onebot>=2.0.0"
]
requires-python = ">=3.8"
readme = "README.md"
license = {text = "MIT"}

[project.urls]
Homepage = "https://github.com/tniany/nonebot_plugin_fishcat_ping"
Repository = "https://github.com/tniany/nonebot_plugin_fishcat_ping"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend" 